refactor(module): move constructStyles to HTMLStyles
